+++ src/main/java/fr/maxlego08/essentials/commands/commands/economy/CommandEconomy.java
@@ -16,6 +16,7 @@ public class CommandEconomy extends VCommand {
         this.setPermission(Permission.ESSENTIALS_ECO_USE);
         this.setDescription(Message.DESCRIPTION_ECO);
         this.addSubCommand(new CommandEconomyGive(plugin));
+        this.addSubCommand(new CommandEconomyGiveRandom(plugin));
         this.addSubCommand(new CommandEconomyGiveAll(plugin));
         this.addSubCommand(new CommandEconomyTake(plugin));
         this.addSubCommand(new CommandEconomySet(plugin));